refactor: registerForPushNotificationsAsync을 system 폴더로 이동
