+++ fourtoon-cookie/src/system/notification.ts
@@ -1,0 +1,46 @@
+import {Platform} from 'react-native';
+import * as Device from 'expo-device';
+import * as Notifications from 'expo-notifications';
+import Constants from 'expo-constants';
+import {OS} from "../types/os";
+
+
+export const registerForPushNotificationsAsync = async (): Promise<string> => {
+    if (Platform.OS === OS.ANDROID) {
+        await Notifications.setNotificationChannelAsync('default', {
+            name: 'default',
+            importance: Notifications.AndroidImportance.MAX,
+            vibrationPattern: [0, 250, 250, 250],
+            lightColor: '#FF231F7C',
+        });
+    }
+
+    if (Device.isDevice) {
+        const {status: existingStatus} = await Notifications.getPermissionsAsync();
+        let finalStatus = existingStatus;
+        if (existingStatus !== 'granted') {
+            const {status} = await Notifications.requestPermissionsAsync();
+            finalStatus = status;
+        }
+        if (finalStatus !== 'granted') {
+            throw new Error('Permission not granted to get push token for push notification!')
+        }
+        const projectId =
+            Constants?.expoConfig?.extra?.eas?.projectId ?? Constants?.easConfig?.projectId;
+        if (!projectId) {
+            throw new Error('Project ID not found');
+        }
+        try {
+            const pushTokenString = (
+                await Notifications.getExpoPushTokenAsync({
+                    projectId,
+                })
+            ).data;
+            return pushTokenString;
+        } catch (e: unknown) {
+            throw new Error(`${e}`)
+        }
+    } else {
+        throw new Error('Must use physical device for push notifications');
+    }
+}
